crdb-postgres/generator/requirements.txt
psycopg2-binary==2.9.5
faker==37.4.0